fix(executor): extract claude.go helpers to sibling file (<800 gate)
Round-2's hook-reality dispatch comment pushed claude.go from 799 to 829,
tripping the CI check-file-sizes gate (>800). Extract the four pure,
receiver-less helpers (getErrorMessage, isCloudWorkspace, isValidUUID,
intFromAny) into helpers.go in the same package — zero behavior change,
claude.go now 785 with real headroom. Caught by Gate 3b (remote-only gate).
